Use Juttle AST visitor to build where clauses
Building of where clauses is a two pass walk through the AST.

In the first pass, more complex juttle constructs unsupported by influxql get
replaced with equivalent, simpler ones. Namely `x in ['a','b','c']` is
replaced with `x = 'a' or x = 'b' or x = 'c'` and unary negation is
propagated into the expression and resolved (TODO).

In the second pass, the actual node values and operators are extracted
and concatenated. A notable exception is the `x =~ "glob"`, which is
translated into an equivalent regular expression.