test: add a hook to override the db host and increase test tolerance
To support running influx inside a container on a mac, update the
live test so that it connects to the hostname in the INFLUX_HOST
environment variable if set, otherwise uses localhost.

Also increase the test retry tolerance so that it waits up to 2 seconds instead
of only 250ms since my setup failed to complete in time when influx is in a
container.